Move database accesses to AuditEntryUtility
