+++ alert-common/src/main/java/com/synopsys/integration/alert/common/data/BaseAuditUtility.java
@@ -25,14 +25,32 @@ package com.synopsys.integration.alert.common.data;
 
 import java.util.Collection;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
+import java.util.function.Function;
 
+import com.synopsys.integration.alert.common.data.model.AlertNotificationWrapper;
+import com.synopsys.integration.alert.common.data.model.AlertPagedModel;
+import com.synopsys.integration.alert.common.data.model.AuditEntryModel;
+import com.synopsys.integration.alert.common.data.model.AuditJobStatusModel;
+import com.synopsys.integration.alert.common.enumeration.AuditEntryStatus;
 import com.synopsys.integration.alert.common.model.AggregateMessageContent;
 
-public interface AuditUtility {
+public interface BaseAuditUtility {
+    Optional<Long> findMatchingAuditId(final Long notificationId, final UUID commonDistributionId);
+
+    Optional<AuditJobStatusModel> findFirstByJobId(final UUID jobId);
+
+    AlertPagedModel<AuditEntryModel> getPageOfAuditEntries(final Integer pageNumber, final Integer pageSize, final String searchTerm, final String sortField, final String sortOrder, final boolean onlyShowSentNotifications,
+        final Function<AlertNotificationWrapper, AuditEntryModel> notificationToAuditEntryConverter);
+
     Map<Long, Long> createAuditEntry(final Map<Long, Long> existingNotificationIdToAuditId, final UUID jobId, final AggregateMessageContent content);
 
     void setAuditEntrySuccess(final Collection<Long> auditEntryIds);
 
     void setAuditEntryFailure(final Collection<Long> auditEntryIds, final String errorMessage, final Throwable t);
+
+    AuditEntryModel convertToAuditEntryModelFromNotification(final AlertNotificationWrapper notificationContentEntry);
+
+    AuditEntryStatus getWorstStatus(final AuditEntryStatus overallStatus, final AuditEntryStatus currentStatus);
 }
