+++ alert-common/src/main/java/com/synopsys/integration/alert/common/data/model/JobAuditModel.java
@@ -21,9 +21,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.web.audit;
-
-import com.synopsys.integration.alert.common.data.model.Config;
+package com.synopsys.integration.alert.common.data.model;
 
 public class JobAuditModel extends Config {
     private String configId;
