+++ alert-database/build.gradle
@@ -1,9 +1,10 @@
 ext.moduleName = 'com.synopsys.integration.alert.database'
 
 dependencies {
-    implementation project(':alert-common')
-
     runtimeOnly 'com.h2database:h2'
+
     api 'org.liquibase:liquibase-core'
     api 'org.springframework.boot:spring-boot-starter-data-jpa'
+    
+    implementation project(':alert-common')
 }
