+++ src/main/java/com/synopsys/integration/alert/channel/slack/SlackChannel.java
@@ -47,7 +47,7 @@ import com.synopsys.integration.alert.common.exception.AlertException;
 import com.synopsys.integration.alert.common.model.AggregateMessageContent;
 import com.synopsys.integration.alert.common.model.CategoryItem;
 import com.synopsys.integration.alert.common.model.LinkableItem;
-import com.synopsys.integration.alert.database.api.AuditEntryAccessor;
+import com.synopsys.integration.alert.database.api.AuditEntryUtility;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.rest.request.Request;
 
@@ -69,7 +69,7 @@ public class SlackChannel extends RestDistributionChannel {
     }
 
     @Autowired
-    public SlackChannel(final Gson gson, final AlertProperties alertProperties, final AuditEntryAccessor auditUtility,
+    public SlackChannel(final Gson gson, final AlertProperties alertProperties, final AuditEntryUtility auditUtility,
         final ChannelRestConnectionFactory channelRestConnectionFactory) {
         super(COMPONENT_NAME, gson, alertProperties, auditUtility, channelRestConnectionFactory);
     }
