+++ src/main/java/com/synopsys/integration/alert/web/audit/AuditEntryActions.java
@@ -23,44 +23,30 @@
  */
 package com.synopsys.integration.alert.web.audit;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.synopsys.integration.alert.channel.ChannelTemplateManager;
+import com.synopsys.integration.alert.common.data.BaseAuditUtility;
 import com.synopsys.integration.alert.common.data.model.AlertNotificationWrapper;
+import com.synopsys.integration.alert.common.data.model.AlertPagedModel;
+import com.synopsys.integration.alert.common.data.model.AuditEntryModel;
+import com.synopsys.integration.alert.common.data.model.AuditJobStatusModel;
 import com.synopsys.integration.alert.common.data.model.CommonDistributionConfiguration;
-import com.synopsys.integration.alert.common.enumeration.AuditEntryStatus;
 import com.synopsys.integration.alert.common.event.DistributionEvent;
 import com.synopsys.integration.alert.common.exception.AlertJobMissingException;
 import com.synopsys.integration.alert.common.exception.AlertNotificationPurgedException;
 import com.synopsys.integration.alert.database.api.JobConfigReader;
 import com.synopsys.integration.alert.database.api.NotificationManager;
-import com.synopsys.integration.alert.database.audit.AuditEntryEntity;
-import com.synopsys.integration.alert.database.audit.AuditEntryRepository;
-import com.synopsys.integration.alert.database.audit.AuditNotificationRelation;
-import com.synopsys.integration.alert.database.audit.AuditNotificationRepository;
-import com.synopsys.integration.alert.database.notification.NotificationContent;
-import com.synopsys.integration.alert.web.model.AlertPagedModel;
-import com.synopsys.integration.alert.web.model.NotificationConfig;
-import com.synopsys.integration.alert.web.model.NotificationContentConverter;
 import com.synopsys.integration.alert.workflow.processor.NotificationProcessor;
 import com.synopsys.integration.exception.IntegrationException;
 
@@ -69,22 +55,18 @@ import com.synopsys.integration.exception.IntegrationException;
 public class AuditEntryActions {
     private final Logger logger = LoggerFactory.getLogger(AuditEntryActions.class);
 
-    private final AuditEntryRepository auditEntryRepository;
-    private final AuditNotificationRepository auditNotificationRepository;
+    private final BaseAuditUtility auditUtility;
     private final NotificationManager notificationManager;
     private final JobConfigReader jobConfigReader;
-    private final NotificationContentConverter notificationContentConverter;
     private final ChannelTemplateManager channelTemplateManager;
     private final NotificationProcessor notificationProcessor;
 
     @Autowired
-    public AuditEntryActions(final AuditEntryRepository auditEntryRepository, final NotificationManager notificationManager, final AuditNotificationRepository auditNotificationRepository,
-        final JobConfigReader jobConfigReader, final NotificationContentConverter notificationContentConverter, final ChannelTemplateManager channelTemplateManager, final NotificationProcessor notificationProcessor) {
-        this.auditEntryRepository = auditEntryRepository;
+    public AuditEntryActions(final BaseAuditUtility auditUtility, final NotificationManager notificationManager, final JobConfigReader jobConfigReader, final ChannelTemplateManager channelTemplateManager,
+        final NotificationProcessor notificationProcessor) {
+        this.auditUtility = auditUtility;
         this.notificationManager = notificationManager;
-        this.auditNotificationRepository = auditNotificationRepository;
         this.jobConfigReader = jobConfigReader;
-        this.notificationContentConverter = notificationContentConverter;
         this.channelTemplateManager = channelTemplateManager;
         this.notificationProcessor = notificationProcessor;
     }
@@ -94,9 +76,7 @@ public class AuditEntryActions {
     }
 
     public AlertPagedModel<AuditEntryModel> get(final Integer pageNumber, final Integer pageSize, final String searchTerm, final String sortField, final String sortOrder, final boolean onlyShowSentNotifications) {
-        final Page<AlertNotificationWrapper> auditPage = queryForNotifications(sortField, sortOrder, searchTerm, pageNumber, pageSize, onlyShowSentNotifications);
-        final List<AuditEntryModel> auditEntries = createRestModels(auditPage.getContent(), sortField, sortOrder);
-        final AlertPagedModel<AuditEntryModel> pagedRestModel = new AlertPagedModel<>(auditPage.getTotalPages(), auditPage.getNumber(), auditEntries.size(), auditEntries);
+        final AlertPagedModel<AuditEntryModel> pagedRestModel = auditUtility.getPageOfAuditEntries(pageNumber, pageSize, searchTerm, sortField, sortOrder, onlyShowSentNotifications, auditUtility::convertToAuditEntryModelFromNotification);
         logger.debug("Paged Audit Entry Rest Model: {}", pagedRestModel);
         return pagedRestModel;
     }
@@ -104,29 +84,14 @@ public class AuditEntryActions {
     public Optional<AuditEntryModel> get(final Long id) {
         if (id != null) {
             final Optional<AlertNotificationWrapper> notificationContent = notificationManager.findById(id);
-            return notificationContent.map(this::createRestModel);
+            return notificationContent.map(auditUtility::convertToAuditEntryModelFromNotification);
         }
         return Optional.empty();
     }
 
     public Optional<AuditJobStatusModel> getAuditInfoForJob(final UUID jobId) {
         if (jobId != null) {
-            final Optional<AuditEntryEntity> auditEntryEntity = auditEntryRepository.findFirstByCommonConfigIdOrderByTimeLastSentDesc(jobId);
-            return auditEntryEntity.map(value -> {
-                String timeCreated = null;
-                if (null != value.getTimeCreated()) {
-                    timeCreated = notificationContentConverter.getContentConverter().getStringValue(value.getTimeCreated());
-                }
-                String timeLastSent = null;
-                if (null != value.getTimeLastSent()) {
-                    timeLastSent = notificationContentConverter.getContentConverter().getStringValue(value.getTimeLastSent());
-                }
-                String status = null;
-                if (null != value.getStatus()) {
-                    status = value.getStatus();
-                }
-                return new AuditJobStatusModel(timeCreated, timeLastSent, status);
-            });
+            return auditUtility.findFirstByJobId(jobId);
         }
         return Optional.empty();
     }
@@ -150,119 +115,12 @@ public class AuditEntryActions {
         }
         distributionEvents.forEach(event -> {
             final UUID commonDistributionId = UUID.fromString(event.getConfigId());
-            Long auditId = null;
-            final Optional<AuditEntryEntity> auditEntryEntity = auditEntryRepository.findMatchingAudit(notificationContent.getId(), commonDistributionId);
+            final Long auditId = auditUtility.findMatchingAuditId(notificationContent.getId(), commonDistributionId).orElse(null);
             final Map<Long, Long> notificationIdToAuditId = new HashMap<>();
-            if (auditEntryEntity.isPresent()) {
-                auditId = auditEntryEntity.get().getId();
-            }
             notificationIdToAuditId.put(notificationContent.getId(), auditId);
             event.setNotificationIdToAuditId(notificationIdToAuditId);
             channelTemplateManager.sendEvent(event);
         });
         return get();
-    }
-
-    private Page<AlertNotificationWrapper> queryForNotifications(final String sortField, final String sortOrder, final String searchTerm, final Integer pageNumber, final Integer pageSize, final boolean onlyShowSentNotifications) {
-        final PageRequest pageRequest = notificationManager.getPageRequestForNotifications(pageNumber, pageSize, sortField, sortOrder);
-        final Page<AlertNotificationWrapper> auditPage;
-        if (StringUtils.isNotBlank(searchTerm)) {
-            auditPage = notificationManager.findAllWithSearch(searchTerm, pageRequest, onlyShowSentNotifications);
-        } else {
-            auditPage = notificationManager.findAll(pageRequest, onlyShowSentNotifications);
-        }
-        return auditPage;
-    }
-
-    private List<AuditEntryModel> createRestModels(final List<AlertNotificationWrapper> notificationContentEntries, final String sortField, final String sortOrder) {
-        final List<AuditEntryModel> auditEntryModels = notificationContentEntries.stream().map(this::createRestModel).collect(Collectors.toList());
-        if (StringUtils.isBlank(sortField) || sortField.equalsIgnoreCase("lastSent") || sortField.equalsIgnoreCase("overallStatus")) {
-            // We do this sorting here because lastSent is not a field in the NotificationContent entity and overallStatus is not stored in the database
-            boolean ascendingOrder = false;
-            if (StringUtils.isNotBlank(sortOrder) && Sort.Direction.ASC.name().equalsIgnoreCase(sortOrder)) {
-                ascendingOrder = true;
-            }
-            Comparator comparator;
-            if (StringUtils.isBlank(sortField) || sortField.equalsIgnoreCase("lastSent")) {
-                final Function<AuditEntryModel, Date> function = auditEntryModel -> {
-                    Date date = null;
-                    if (StringUtils.isNotBlank(auditEntryModel.getLastSent())) {
-                        date = notificationContentConverter.parseDateString(auditEntryModel.getLastSent());
-                    }
-                    return date;
-                };
-                comparator = Comparator.comparing(function, Comparator.nullsFirst(Comparator.reverseOrder()));
-            } else {
-                comparator = Comparator.comparing(AuditEntryModel::getOverallStatus, Comparator.nullsLast(String::compareTo));
-            }
-            if (ascendingOrder) {
-                comparator = comparator.reversed();
-            }
-            auditEntryModels.sort(comparator);
-        }
-        return auditEntryModels;
-    }
-
-    private AuditEntryModel createRestModel(final AlertNotificationWrapper notificationContentEntry) {
-        final List<AuditNotificationRelation> relations = auditNotificationRepository.findByNotificationId(notificationContentEntry.getId());
-        final List<Long> auditEntryIds = relations.stream().map(AuditNotificationRelation::getAuditEntryId).collect(Collectors.toList());
-        final List<AuditEntryEntity> auditEntryEntities = auditEntryRepository.findAllById(auditEntryIds);
-
-        AuditEntryStatus overallStatus = null;
-        String lastSent = null;
-        Date lastSentDate = null;
-        final List<JobAuditModel> jobAuditModels = new ArrayList<>();
-        for (final AuditEntryEntity auditEntryEntity : auditEntryEntities) {
-            final UUID commonConfigId = auditEntryEntity.getCommonConfigId();
-
-            if (null == lastSentDate || (null != auditEntryEntity.getTimeLastSent() && lastSentDate.before(auditEntryEntity.getTimeLastSent()))) {
-                lastSentDate = auditEntryEntity.getTimeLastSent();
-                lastSent = notificationContentConverter.getContentConverter().getStringValue(lastSentDate);
-            }
-            final String id = notificationContentConverter.getContentConverter().getStringValue(auditEntryEntity.getId());
-            final String configId = notificationContentConverter.getContentConverter().getStringValue(commonConfigId);
-            final String timeCreated = notificationContentConverter.getContentConverter().getStringValue(auditEntryEntity.getTimeCreated());
-            final String timeLastSent = notificationContentConverter.getContentConverter().getStringValue(auditEntryEntity.getTimeLastSent());
-
-            AuditEntryStatus status = null;
-            if (auditEntryEntity.getStatus() != null) {
-                status = AuditEntryStatus.valueOf(auditEntryEntity.getStatus());
-                overallStatus = getWorstStatus(overallStatus, status);
-            }
-
-            final String errorMessage = auditEntryEntity.getErrorMessage();
-            final String errorStackTrace = auditEntryEntity.getErrorStackTrace();
-
-            final Optional<CommonDistributionConfiguration> commonConfig = jobConfigReader.getPopulatedJobConfig(commonConfigId);
-            String distributionConfigName = null;
-            String eventType = null;
-            if (commonConfig.isPresent()) {
-                distributionConfigName = commonConfig.get().getName();
-                eventType = commonConfig.get().getChannelName();
-            }
-
-            String statusDisplayName = null;
-            if (null != status) {
-                statusDisplayName = status.getDisplayName();
-            }
-            final AuditJobStatusModel auditJobStatusModel = new AuditJobStatusModel(timeCreated, timeLastSent, statusDisplayName);
-            jobAuditModels.add(new JobAuditModel(id, configId, distributionConfigName, eventType, auditJobStatusModel, errorMessage, errorStackTrace));
-        }
-        final String id = notificationContentConverter.getContentConverter().getStringValue(notificationContentEntry.getId());
-        final NotificationConfig notificationConfig = (NotificationConfig) notificationContentConverter.populateConfigFromEntity((NotificationContent) notificationContentEntry);
-
-        String overallStatusDisplayName = null;
-        if (null != overallStatus) {
-            overallStatusDisplayName = overallStatus.getDisplayName();
-        }
-        return new AuditEntryModel(id, notificationConfig, jobAuditModels, overallStatusDisplayName, lastSent);
-    }
-
-    private AuditEntryStatus getWorstStatus(final AuditEntryStatus overallStatus, final AuditEntryStatus currentStatus) {
-        AuditEntryStatus newOverallStatus = overallStatus;
-        if (null == overallStatus || currentStatus == AuditEntryStatus.FAILURE || (AuditEntryStatus.SUCCESS == overallStatus && AuditEntryStatus.SUCCESS != currentStatus)) {
-            newOverallStatus = currentStatus;
-        }
-        return newOverallStatus;
     }
 }
