+++ src/test/java/com/synopsys/integration/alert/audit/mock/MockAuditEntryRestModel.java
@@ -9,11 +9,11 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
+import com.synopsys.integration.alert.common.data.model.AuditEntryModel;
+import com.synopsys.integration.alert.common.data.model.JobAuditModel;
+import com.synopsys.integration.alert.common.data.model.NotificationConfig;
 import com.synopsys.integration.alert.common.enumeration.AuditEntryStatus;
 import com.synopsys.integration.alert.mock.model.MockRestModelUtil;
-import com.synopsys.integration.alert.web.audit.AuditEntryModel;
-import com.synopsys.integration.alert.web.audit.JobAuditModel;
-import com.synopsys.integration.alert.web.model.NotificationConfig;
 
 public class MockAuditEntryRestModel extends MockRestModelUtil<AuditEntryModel> {
     private final String timeLastSent = new Date(500).toString();
