+++ src/test/java/com/synopsys/integration/alert/channel/ChannelTemplateManagerTest.java
@@ -19,14 +19,14 @@ import com.synopsys.integration.alert.common.event.ContentEvent;
 import com.synopsys.integration.alert.common.event.DistributionEvent;
 import com.synopsys.integration.alert.common.model.AggregateMessageContent;
 import com.synopsys.integration.alert.common.model.LinkableItem;
-import com.synopsys.integration.alert.database.api.AuditEntryAccessor;
+import com.synopsys.integration.alert.database.api.AuditEntryUtility;
 import com.synopsys.integration.rest.RestConstants;
 
 public class ChannelTemplateManagerTest {
 
     @Test
     public void testSendEvents() {
-        final AuditEntryAccessor auditUtility = Mockito.mock(AuditEntryAccessor.class);
+        final AuditEntryUtility auditUtility = Mockito.mock(AuditEntryUtility.class);
         final JmsTemplate jmsTemplate = Mockito.mock(JmsTemplate.class);
         final ContentConverter contentConverter = Mockito.mock(ContentConverter.class);
         Mockito.doNothing().when(jmsTemplate).convertAndSend(Mockito.anyString(), Mockito.any(Object.class));
@@ -42,7 +42,7 @@ public class ChannelTemplateManagerTest {
 
     @Test
     public void testNotAbstractChannelEvent() {
-        final AuditEntryAccessor auditUtility = Mockito.mock(AuditEntryAccessor.class);
+        final AuditEntryUtility auditUtility = Mockito.mock(AuditEntryUtility.class);
         final JmsTemplate jmsTemplate = Mockito.mock(JmsTemplate.class);
         final ContentConverter contentConverter = Mockito.mock(ContentConverter.class);
         Mockito.doNothing().when(jmsTemplate).convertAndSend(Mockito.anyString(), Mockito.any(Object.class));
