+++ src/test/java/com/synopsys/integration/alert/channel/hipchat/HipChatChannelTest.java
@@ -36,7 +36,7 @@ import com.synopsys.integration.alert.common.enumeration.FormatType;
 import com.synopsys.integration.alert.common.event.DistributionEvent;
 import com.synopsys.integration.alert.common.model.AggregateMessageContent;
 import com.synopsys.integration.alert.common.model.LinkableItem;
-import com.synopsys.integration.alert.database.api.AuditEntryAccessor;
+import com.synopsys.integration.alert.database.api.AuditEntryUtility;
 import com.synopsys.integration.alert.provider.blackduck.BlackDuckProvider;
 import com.synopsys.integration.alert.util.TestAlertProperties;
 import com.synopsys.integration.alert.util.TestPropertyKey;
@@ -53,7 +53,7 @@ public class HipChatChannelTest extends ChannelTest {
     @Tag(TestTags.CUSTOM_EXTERNAL_CONNECTION)
     @Disabled("Hip Chat public api is currently end of life; need an on premise installation to test")
     public void sendMessageTestIT() throws IOException, IntegrationException {
-        final AuditEntryAccessor auditUtility = Mockito.mock(AuditEntryAccessor.class);
+        final AuditEntryUtility auditUtility = Mockito.mock(AuditEntryUtility.class);
         final TestAlertProperties testAlertProperties = new TestAlertProperties();
         final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
         Mockito.when(proxyManager.createProxyInfo()).thenReturn(ProxyInfo.NO_PROXY_INFO);
@@ -100,7 +100,7 @@ public class HipChatChannelTest extends ChannelTest {
 
     @Test
     public void createRequestThrowsExceptionForTemplateTest() throws Exception {
-        final AuditEntryAccessor auditUtility = Mockito.mock(AuditEntryAccessor.class);
+        final AuditEntryUtility auditUtility = Mockito.mock(AuditEntryUtility.class);
         final HipChatChannel hipChatChannel = new HipChatChannel(gson, null, auditUtility, null);
 
         final LinkableItem subTopic = new LinkableItem("subTopic", "Alert has sent this test message", null);
@@ -131,7 +131,7 @@ public class HipChatChannelTest extends ChannelTest {
 
     @Test
     public void testEmptyContent() throws Exception {
-        final AuditEntryAccessor auditUtility = Mockito.mock(AuditEntryAccessor.class);
+        final AuditEntryUtility auditUtility = Mockito.mock(AuditEntryUtility.class);
         final TestAlertProperties alertProperties = new TestAlertProperties();
         final HipChatChannel hipChatChannel = new HipChatChannel(gson, alertProperties, auditUtility, null);
 
@@ -153,7 +153,7 @@ public class HipChatChannelTest extends ChannelTest {
 
     @Test
     public void testChunkedRequestList() throws Exception {
-        final AuditEntryAccessor auditUtility = Mockito.mock(AuditEntryAccessor.class);
+        final AuditEntryUtility auditUtility = Mockito.mock(AuditEntryUtility.class);
         final TestAlertProperties alertProperties = new TestAlertProperties();
         final HipChatChannel hipChatChannel = new HipChatChannel(gson, alertProperties, auditUtility, null);
