+++ src/test/java/com/synopsys/integration/alert/mock/model/MockNotificationRestModel.java
@@ -3,7 +3,7 @@ package com.synopsys.integration.alert.mock.model;
 import java.util.Date;
 
 import com.google.gson.JsonObject;
-import com.synopsys.integration.alert.web.model.NotificationConfig;
+import com.synopsys.integration.alert.common.data.model.NotificationConfig;
 import com.synopsys.integration.blackduck.api.generated.enumeration.NotificationType;
 
 public class MockNotificationRestModel extends MockRestModelUtil<NotificationConfig> {
